requirements/local.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/local.txt requirements/local.in
#
asgiref==3.8.1
    # via
    #   -r requirements/local.in
    #   django
    #   django-cors-headers
django==5.1.3
    # via
    #   -r requirements/local.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==4.6.0
    # via -r requirements/local.in
djangorestframework==3.15.2
    # via
    #   -r requirements/local.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via -r requirements/local.in
drf-yasg==1.21.8
    # via -r requirements/local.in
inflection==0.5.1
    # via
    #   -r requirements/local.in
    #   drf-yasg
iniconfig==2.0.0
    # via
    #   -r requirements/local.in
    #   pytest
packaging==24.2
    # via
    #   -r requirements/local.in
    #   drf-yasg
    #   pytest
pluggy==1.5.0
    # via
    #   -r requirements/local.in
    #   pytest
psycopg2==2.9.10
    # via -r requirements/local.in
pyjwt==2.9.0
    # via
    #   -r requirements/local.in
    #   djangorestframework-simplejwt
pytest==8.3.3
    # via -r requirements/local.in
pytz==2024.2
    # via
    #   -r requirements/local.in
    #   drf-yasg
pyyaml==6.0.2
    # via
    #   -r requirements/local.in
    #   drf-yasg
sqlparse==0.5.1
    # via
    #   -r requirements/local.in
    #   django
uritemplate==4.1.1
    # via
    #   -r requirements/local.in
    #   drf-yasg
